clean output before the simulation
