+++ tests/end_to_end/utils_functions.py
@@ -1,5 +1,6 @@
 import json
 import os
+import shutil
 import sys
 import zipfile
 from pathlib import Path
@@ -42,6 +43,13 @@ def get_mpi_command(allow_run_as_root=False, nproc: int = 1):
         else:
             return [MPI_LAUNCHER, MPI_N, nproc_str, "--oversubscribe"]
 
+
+def remove_outputs(study_path):
+    output_path = study_path / "output"
+    if os.path.isdir(output_path):
+        for f in Path(output_path).iterdir():
+            if f.is_dir():
+                shutil.rmtree(f)
 
 def get_filepath(output_dir, folder, filename):
     op = []
